Remove namespace unaware rotation methods
